+++ k9mail-library/src/main/java/com/fsck/k9/mail/internet/MessageExtractor.java
@@ -26,7 +26,7 @@ import org.apache.commons.io.input.BoundedInputStream;
 import static com.fsck.k9.mail.K9MailLib.LOG_TAG;
 import static com.fsck.k9.mail.internet.CharsetSupport.fixupCharset;
 import static com.fsck.k9.mail.internet.MimeUtility.getHeaderParameter;
-import static com.fsck.k9.mail.internet.MimeUtility.isFormatFlowed;
+import static com.fsck.k9.mail.internet.FlowedMessageUtils.isFormatFlowed;
 import static com.fsck.k9.mail.internet.MimeUtility.isSameMimeType;
 import static com.fsck.k9.mail.internet.Viewable.Alternative;
 import static com.fsck.k9.mail.internet.Viewable.Html;
@@ -193,7 +193,8 @@ public class MessageExtractor {
             Viewable viewable;
             if (isSameMimeType(mimeType, "text/plain")) {
                 if (isFormatFlowed(part.getContentType())) {
-                    viewable = new Flowed(part);
+                    boolean delSp = FlowedMessageUtils.isDelSp(part.getContentType());
+                    viewable = new Flowed(part, delSp);
                 } else {
                     viewable = new Text(part);
                 }
