+++ k9mail-library/src/main/java/com/fsck/k9/mail/internet/MimeUtility.java
@@ -28,9 +28,6 @@ import static com.fsck.k9.mail.K9MailLib.LOG_TAG;
 public class MimeUtility {
     public static final String DEFAULT_ATTACHMENT_MIME_TYPE = "application/octet-stream";
     public static final String K9_SETTINGS_MIME_TYPE = "application/x-k9settings";
-    private static final String TEXT_PLAIN = "text/plain";
-    private static final String HEADER_PARAM_FORMAT = "format";
-    private static final String HEADER_FORMAT_FLOWED = "flowed";
 
     /*
      * http://www.w3schools.com/media/media_mimeref.asp
@@ -1143,14 +1140,5 @@ public class MimeUtility {
 
     public static boolean isSameMimeType(String mimeType, String otherMimeType) {
         return mimeType != null && mimeType.equalsIgnoreCase(otherMimeType);
-    }
-
-    static boolean isFormatFlowed(String contentType) {
-        String mimeType = getHeaderParameter(contentType, null);
-        if (isSameMimeType(TEXT_PLAIN, mimeType)) {
-            String formatParameter = getHeaderParameter(contentType, HEADER_PARAM_FORMAT);
-            return HEADER_FORMAT_FLOWED.equalsIgnoreCase(formatParameter);
-        }
-        return false;
     }
 }
