+++ k9mail-library/src/test/java/com/fsck/k9/mail/internet/MimeUtilityTest.java
@@ -133,21 +133,4 @@ public class MimeUtilityTest {
     public void isSameMimeType_withSecondArgumentBeingNull_shouldReturnFalse() throws Exception {
         assertFalse(MimeUtility.isSameMimeType("text/html", null));
     }
-
-    @Test
-    public void isFormatFlowed_withTextPlainFormatFlowed__shouldReturnTrue() throws Exception {
-        assertTrue(MimeUtility.isFormatFlowed("text/plain; format=flowed"));
-    }
-
-    @Test
-    public void isFormatFlowed_withTextPlain__shouldReturnFalse() throws Exception {
-        assertFalse(MimeUtility.isFormatFlowed("text/plain"));
-    }
-
-    @Test
-    public void isFormatFlowed_withTextHtmlFormatFlowed__shouldReturnFalse() throws Exception {
-        assertFalse(MimeUtility.isFormatFlowed("text/html; format=flowed"));
-    }
-
-
 }
